requirements.txt
tweepy==4.6.0
python-dotenv==0.20.0
pandas==1.4.2
transformers==4.18.0
datasets==2.1.0
flask==2.1.2
instaloader==4.9
requests
requests-oauthlib
werkzeug==2.0.3
